server/stocks/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.12.2
better-profanity==0.7.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
Django==5.0.2
django-celery-beat==2.7.0
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
filelock==3.18.0
fsspec==2025.3.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
kombu==5.5.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.1
packaging==24.2
paho-mqtt==1.6.1
pillow==10.2.0
prompt_toolkit==3.0.50
PyJWT==2.6.0
pyngrok==7.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.2
redis==5.2.1
requests==2.31.0
schedule==1.2.0
six==1.17.0
soupsieve==2.6
sqlparse==0.4.3
sympy==1.13.1
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
djangorestframework-simplejwt==5.2.2
Pillow==10.2.0
# websocket-client==1.7.0
